python-fast-api/requirements.txt
Faker==19.11.0
fastapi==0.103.2
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
psycopg2-binary==2.9.7
PyJWT==2.3.0
XlsxWriter==3.1.2
pkgutil==
